Move EdgeData, VertexData, Protocols and ConnStates to a new model package.
