+++ csb/src/main/scala/edu/msstate/dasi/csb/GraphSynth.scala
@@ -1,6 +1,7 @@
 package edu.msstate.dasi.csb
 
 import edu.msstate.dasi.csb.distributions.DataDistributions
+import edu.msstate.dasi.csb.model.{EdgeData, VertexData}
 import org.apache.spark.graphx.Graph
 
 /**
