+++ csb/src/main/scala/edu/msstate/dasi/csb/KroFit.scala
@@ -1,5 +1,6 @@
 package edu.msstate.dasi.csb
 
+import edu.msstate.dasi.csb.model.{EdgeData, VertexData}
 import org.apache.spark.SparkContext
 import org.apache.spark.graphx.Graph
 import org.apache.spark.rdd.RDD
