--- csb/src/main/scala/edu/msstate/dasi/csb/VertexData.scala
@@ -1,19 +1,0 @@
-package edu.msstate.dasi.csb
-
-/**
- *
- */
-case class VertexData()
-
-object VertexData {
-  /**
-   *
-   */
-  def apply(text: String): VertexData = {
-    if (text == "null") {
-      null.asInstanceOf[VertexData]
-    } else {
-      new VertexData()
-    }
-  }
-}
