+++ csb/src/main/scala/edu/msstate/dasi/csb/model/Protocols.scala
@@ -1,0 +1,13 @@
+package edu.msstate.dasi.csb.model
+
+object Protocols extends Enumeration {
+  /**
+    * transport_proto  Meaning
+    * UNKNOWN  An unknown transport-layer protocol.
+    * TCP  Transmission Control Protocol
+    * UDP  User Datagram Protocol
+    * ICMP  Internet Control Message Protocol
+    *
+    */
+  val UNKNOWN, TCP, UDP, ICMP = Value
+}
