+++ csb/src/main/scala/edu/msstate/dasi/csb/model/VertexData.scala
@@ -1,0 +1,19 @@
+package edu.msstate.dasi.csb.model
+
+/**
+ *
+ */
+case class VertexData()
+
+object VertexData {
+  /**
+   *
+   */
+  def apply(text: String): VertexData = {
+    if (text == "null") {
+      null.asInstanceOf[VertexData]
+    } else {
+      new VertexData()
+    }
+  }
+}
